# Detect bad unstructured conversions