@wojtek-t 2017-02-14
@smarterclayton - It's still not fully ready for review, but I fixed a bunch of issues. And at the same time, this is faster now (and less allocations) - I updated the PR description.