@smarterclayton 2017-02-17
Two comments to have a consistent setup / constructor.  Everything else looks fine.