@wojtek-t 2017-02-17
Fixed gofmt.