@smarterclayton 2017-02-17
This flake link
<https://k8s-gubernator.appspot.com/build/kubernetes-jenkins/pr-logs/pull/41401/pull-kubernetes-node-e2e-non-cri/164/>
looks
non-related but I see no other references to it.

@k8s-bot non-cri node e2e test this

On Fri, Feb 17, 2017 at 4:07 PM, k8s-ci-robot <notifications@github.com>
wrote:

> @wojtek-t <https://github.com/wojtek-t>: The following test(s) *failed*:
> Test name Commit Details Rerun command
> Jenkins non-CRI GCE Node e2e 924e9a5
> <https://github.com/kubernetes/kubernetes/commit/924e9a5b3aad8df9b162e3acd2ef0dd2fbf2ffce>
> link
> <https://k8s-gubernator.appspot.com/build/kubernetes-jenkins/pr-logs/pull/41401/pull-kubernetes-node-e2e-non-cri/164/> @k8s-bot
> non-cri node e2e test this
>
> Full PR test history <https://k8s-gubernator.appspot.com/pr/41401>. Your
> PR dashboard <https://k8s-gubernator.appspot.com/pr/wojtek-t>. Please
> help us cut down on flakes by linking to an open issue
> <https://github.com/kubernetes/kubernetes/issues?q=is:issue+is:open> when
> you hit one in your PR.
>
> Instructions for interacting with me using PR comments are available here
> <https://github.com/kubernetes/community/blob/master/contributors/devel/pull-request-commands.md>.
> If you have questions or suggestions related to my behavior, please file an
> issue against the kubernetes/test-infra
> <https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:>
> repository. I understand the commands that are listed here
> <https://github.com/kubernetes/test-infra/blob/master/prow/commands.md>.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/41401#issuecomment-280766588>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABG_p1f_NOKOBX4KHNEAK6Zv20eKycljks5rdgwDgaJpZM4MATJi>
> .
>